styles: grants page ui
